--- config/global-teardown.js
@@ -1,7 +1,0 @@
-const teardown = require("jest-playwright-preset/teardown")
-
-module.exports = async (config) => {
-  await teardown(config)
-
-  global.server.close()
-}
