+++ tsconfig.json
@@ -11,5 +11,5 @@
     "./global.d.ts",
     "node_modules/jest-playwright-preset/types/global.d.ts"
   ],
-  "include": ["src/**/*"]
+  "include": ["config/**/*", "src/**/*"]
 }
